group all constant and type definitions in header .jl files
