+++ bioinformatics/QpcrAnalysis/allelic_discrimination.jl
@@ -4,47 +4,10 @@ import DataStructures.OrderedDict
 import Clustering: ClusteringResult, kmeans!, kmedoids!, silhouettes
 import Combinatorics.combinations
 
-# 4 groups
-const DEFAULT_encgr = Array{Int,2}(0, 0)
-# const DEFAULT_encgr = [0 1 0 1; 0 0 1 1] # NTC, homo ch1, homo ch2, hetero
-const DEFAULT_init_FACTORS = [1, 1, 1, 1] # sometimes "hetero" may not have very high end-point fluo
-const DEFAULT_apg_LABELS = ["ntc", "homo_1", "homo_2", "hetero", "unclassified"] # [0 1 0 1; 0 0 1 1]
-# const DEFAULT_apg_LABELS = ["hetero", "homo_2", "homo_1", "ntc", "unclassified"] # [1 0 1 0; 1 1 0 0]
-
-# # 3 groups without NTC
-# const DEFAULT_egr = [1 0 1; 0 1 1] # homo ch1, homo ch2, hetero
-# const DEFAULT_init_FACTORS = [1, 1, 1] # sometimes "hetero" may not have very high end-point fluo
-# const DEFAULT_eg_LABELS = ["homo_a", "homo_b", "hetero", "unclassified"]
-
-const CATEG_WELL_VEC = [
-    ("rbbs_ary3",   Colon()),
-    ("blsub_fluos", Colon()),
-    ("d0",          Colon()),
-    ("cq",          Colon())
-]
-
-# const CTRL_WELL_VEC = fill(Vector{Int}(), length(DEFAULT_init_FACTORS)) # All empty. NTC, homo ch1, homo ch2, hetero
-const CTRL_WELL_DICT = OrderedDict{Vector{Int},Vector{Int}}() # key is genotype (Vector{Int}), value is well numbers (Vector{Int})
-# # example
-# const CTRL_WELL_DICT = OrderedDict(
-#     [0, 0] => [1, 2], # NTC, well 1 and 2
-#     [1, 0] => [3, 4], # homo ch1, well 3 and 4
-#     [0, 1] => [5, 6], # homo ch2, well 5 and 6
-#     [1, 1] => [7, 8]  # hetero, well 7 and 8
-# )
-# # old approach
-# const CTRL_WELL_DICT = DefaultOrderedDict(Vector{Int}, Vector{Int}, Vector{Int}())
 
 # nrn: whether to flip the binary genotype or not
 NRN_SELF = x -> x
 NRN_NOT = x -> 1 .- x
-
-# # may be needed if not always called by `process_amp_1sr`
-# type AllelicDiscriminationResult #
-#     cluster_result ::ClusteringResult
-#     # assignments_adj ::Vector{Int}
-#     assignments_adj_labels ::Vector{String}
-# end # type
 
 
 # start with bottom-level function, goes step-wise
