+++ bioinformatics/QpcrAnalysis/meltcrv.jl
@@ -4,44 +4,6 @@ import DataStructures.OrderedDict
 import DataArrays.DataArray
 import StatsBase.countmap
 
-const EMPTY_mc = zeros(1,3)[1:0,:]
-const EMPTY_Ta = zeros(1,2)[1:0,:]
-const EMPTY_mc_tm_pw_out = OrderedDict(
-    "mc" => EMPTY_mc,
-    "Ta_raw" => EMPTY_Ta,
-    "Ta_fltd" => EMPTY_Ta
-)
-
-
-struct MeltCurveTF # temperature and fluorescence
-    t_da_vec ::Vector{DataArray{Float64,1}}
-    fluo_da ::DataArray{Float64,2}
-end
-
-struct MeltCurveTa # Tm and area
-    mc ::Array{Float64,2}
-    Ta_fltd ::Array{Float64,2}
-    mc_denser ::Array{Float64,2}
-    ns_range_mid ::Real
-    sn_dict ::OrderedDict{String,Array{Float64,2}}
-    Ta_raw ::Array{Float64,2}
-    Ta_reported ::String
-end
-
-struct MeltCurveOutput
-    mc_bychwl ::Matrix{MeltCurveTa} # dim1 is well and dim2 is channel
-    channel_nums ::Vector{Int}
-    fluo_well_nums ::Vector{Int}
-    fr_ary3 ::Array{Float64,3}
-    mw_ary3 ::Array{Float64,3}
-    k4dcv ::K4Deconv
-    fdcvd_ary3 ::Array{Float64,3}
-    wva_data ::OrderedDict{String,OrderedDict{Int,Vector{Float64}}}
-    wva_well_nums ::Vector{Int}
-    faw_ary3 ::Array{Float64,3}
-    tf_bychwl ::OrderedDict{Int,Vector{OrderedDict{String,Vector{Float64}}}}
-end
-
 
 # called by QpcrAnalyze.dispatch
 function act(
