+++ bioinformatics/QpcrAnalysis/multi_channel.jl
@@ -1,4 +1,6 @@
-# process multi-channel (mtch) data. Use global variable `channels`.
+# multi-channel.jl
+#
+# process multi-channel (mtch) data. 
 
 # Top-level function: run qPCR functions on multi-channel fluorescence data and output consolidated results. original matrix will become 3-D array, original string or list will be come list.
 function process_mtch(
