+++ bioinformatics/test/test_functions.jl
@@ -10,9 +10,9 @@ import DataFrames: DataFrame, rename
 import DataStructures: OrderedDict
 import QpcrAnalysis: dispatch, act, verify_request, verify_response, print_v
 
-td  = readdlm("../test/data/test_data.csv",',',header=true)
+td  = readdlm("$LOAD_FROM_DIR/../test/data/test_data.csv",',',header=true)
 td1 = DataFrame(td[1])
-global const TEST_DATA = rename(td1,zip(names(td1),map(x->Symbol(x),squeeze(td[2],1))))
+const TEST_DATA = rename(td1,zip(names(td1),map(x->Symbol(x),squeeze(td[2],1))))
 
 function test_dispatch(;
     debug     ::Bool =false,
